+++ src/menu_items.c
@@ -525,6 +525,7 @@ char* gTextPracticeMenu[] = {
     "ITEM CONTROL",
     "LAP SKIP (L+UP)",
     "TIMER FREEZE",
+    "SAVE GHOST",
     "RETURN",
 };
 
@@ -536,6 +537,7 @@ char* gTextPracticeHelp[] = {
     "SETS FORCED ITEM (PRACTICE).",
     "ENABLES L+D-PAD UP TO SKIP LAP.",
     "STOPS THE RACE TIMER.",
+    "SAVES CURRENT RACE TO PC DISK.",
     "RETURN TO EXTRAS MENU.",
 };
 
@@ -6980,6 +6982,7 @@ void func_800A1FB0(MenuItem* arg0) {
         case SUB_MENU_PRACTICE_ITEM_OPTION:
         case SUB_MENU_PRACTICE_LAP_SKIP:
         case SUB_MENU_PRACTICE_TIMER_FREEZE:
+        case SUB_MENU_PRACTICE_SAVE_GHOST:
         case SUB_MENU_PRACTICE_RETURN:
             for (i = 0; i < ARRAY_COUNT(gTextPracticeMenu); i++) {
                 set_text_color_rainbow_if_selected(gSubMenuSelection - SUB_MENU_PRACTICE_MIN, i, 3);
@@ -6997,6 +7000,7 @@ void func_800A1FB0(MenuItem* arg0) {
             print_text1_center_mode_1(0xE0, 0x30 + 0x60, gTextItemOptions[gPracticeItemOption], 0, 0.9f, 1.0f);
             print_text1_center_mode_1(0xE0, 0x30 + 0x78, gEnableLapSkip ? gTextOn : gTextOff, 0, 0.9f, 1.0f);
             print_text1_center_mode_1(0xE0, 0x30 + 0x90, gPracticeTimerFreeze ? gTextOn : gTextOff, 0, 0.9f, 1.0f);
+            print_text1_center_mode_1(0xE0, 0x30 + 0xA8, "SAVE TO DISK", 0, 0.9f, 1.0f);
 
             // Draw Description/Tooltip
             set_text_color(TEXT_YELLOW);